Fix pre-existing residuals from the v1.3.2..HEAD review
- localizedVarName advances a numeric suffix from the incoming value's
  root identifier (t -> t2 -> t3): a pointer TypeWrapper wrapping a
  pointer value localizes twice in one brace scope, which previously
  emitted a colliding redeclaration the generator exited 0 on
- the reflection decode of dynamic-element lists and vectors bounds the
  trailing element in int64: uint32(sszLen) wrapped for regions past
  4GiB, rejecting valid input the generated code decodes (element
  starts remain 4-byte offsets, only the trailing extent can exceed)
- the recursion fixup only marks members whose flag is unset, so a
  shared already-published descriptor is not written concurrently with
  lock-free readers
- the container marshal accumulates dynamic-field offsets in int64 so
  the MaxUint32 guard cannot be bypassed by 32-bit int wraparound

+++ reflection/marshal.go
@@ -362,7 +362,7 @@ func (ctx *ReflectionCtx) marshalContainer(sourceType *ssztypes.TypeDescriptor, 
 	}
 
 	offset := 0
-	dynObjOffset := 0
+	dynObjOffset := int64(0)
 	canSeek := encoder.Seekable()
 	startLen := encoder.GetPosition()
 
@@ -395,12 +395,12 @@ func (ctx *ReflectionCtx) marshalContainer(sourceType *ssztypes.TypeDescriptor, 
 					return sszutils.ErrorWithPathf(err, "%s:o", field.Name)
 				}
 
-				fieldOffset := sourceType.Len + int64(dynObjOffset)
+				fieldOffset := sourceType.Len + dynObjOffset
 				if fieldOffset > math.MaxUint32 {
 					return sszutils.ErrOffsetOverflowFn(fieldOffset)
 				}
 				encoder.EncodeOffset(uint32(fieldOffset))
-				dynObjOffset += int(size)
+				dynObjOffset += size
 			}
 			// fmt.Printf("%sfield %d:\t offset [%v:%v] %v\t %v\n", strings.Repeat(" ", int(depth.idt)*2+1), i, offset, offset+fieldSize, fieldSize, field.Name)
 		}
